Ad Clicks A/B Testing
A/B testing is a method used to compare two versions of something to determine which one performs better. The dataset contains information on user interactions with different versions of an advertisement. The goal is to analyze the data to see if there is a statistically significant difference in user behavior between the two versions (A and B).